3 classes and fetched data
I’ve managed to fetch some data from the Facebook API and simply
present this as out.prints(). However, getting friends data is not
possible unless they are members of the same app and will be something
we’ll work around. Now its all about getting the data thats available
and write it to an ref file.